[Project 4-2] Create inumber syscall
