+++ include/filesys/inode.h
@@ -5,7 +5,30 @@
 #include "filesys/off_t.h"
 #include "devices/disk.h"
 
+#include <list.h>
+
 struct bitmap;
+
+/* On-disk inode. //
+ * Must be exactly DISK_SECTOR_SIZE bytes long. */
+struct inode_disk {
+	disk_sector_t start;                /* First data sector. */
+	off_t length;                       /* File size in bytes. */
+	unsigned magic;                     /* Magic number. */
+	//project 4-2
+	bool isdir;
+	uint8_t unused [499];				/* Not used. */
+};
+
+/* In-memory inode. */
+struct inode {
+	struct list_elem elem;              /* Element in inode list. */
+	disk_sector_t sector;               /* Sector number of disk location. */
+	int open_cnt;                       /* Number of openers. */
+	bool removed;                       /* True if deleted, false otherwise. */
+	int deny_write_cnt;                 /* 0: writes ok, >0: deny writes. */
+	struct inode_disk data;             /* Inode content. */
+};
 
 void inode_init (void);
 bool inode_create (disk_sector_t, off_t, bool);
